// src/enum/mensagem.enum.ts
export enum MensagemEnum {
    OCORREU_ERRO = "OCORREU_ERRO",
    OPERACAO_NAO_AUTORIZADA = "OPERACAO_NAO_AUTORIZADA",

    NADA_ENCONTRADO_SUCESSO = "NADA_ENCONTRADO_SUCESSO",
    NADA_ENCONTRADO_ERRO = "NADA_ENCONTRADO_ERRO",
    
    CRIADO_ERRO = "CRIADO_ERRO",
    
    DELETADO_ERRO = "DELETADO_ERRO",
    
    ALTERACOES_ERRO = "ALTERACOES_ERRO",
    
    CAMPO_INVALIDO = "CAMPO_INVALIDO",
    CAMPO_OBRIGATORIO = "CAMPO_OBRIGATORIO",
    
    EMAIL_INVALIDO = "EMAIL_INVALIDO",
    EMAIL_OU_SENHA_INVALIDO = "EMAIL_OU_SENHA_INVALIDO",
    EMAIL_EXISTE = "EMAIL_EXISTE",
    
    SENHA_VALIDA = "SENHA_VALIDA",
};